Agregando librerias

+++ js/main.js
@@ -19,18 +19,18 @@ let inventario = [
   }
 ];
 
-let total = 0
+let total = 0;
 
 let carrito = [];
 
-const divMain = document.getElementById ('container')
-const divCarrito = document.getElementById('carrito')
-const divTotalCarrito = document.getElementById('totalCarrito')
+const divMain = document.getElementById ('container');
+const divCarrito = document.getElementById('carrito');
+const divTotalCarrito = document.getElementById('totalCarrito');
 
 //TARJETA DE PRODUCTOS EN EL HTML
 
 inventario.forEach ((elem) => {
-  const divProductos = document.createElement('div')
+  const divProductos = document.createElement('div');
   divProductos.classList.add("card", "col-sm-6", "col-lg-4", "text-bg-white", "border-light");
   divProductos.innerHTML = `
     <img src="${elem.imagen}" class="card-img-top" alt="...">
@@ -39,31 +39,31 @@ inventario.forEach ((elem) => {
       <p class="card-text">$${elem.precio}.-</p>
       <button href="#" class="btn btn-primary" id="agregar${elem.id}">Agregar al carrito</button>
     </div>`
-  divMain.appendChild(divProductos)
+  divMain.appendChild(divProductos);
 
-  const boton = document.getElementById(`agregar${elem.id}`)
+  const boton = document.getElementById(`agregar${elem.id}`);
 
   boton.addEventListener("click", () => {
-    buscarProducto(elem.id)
-  })
-})
+    buscarProducto(elem.id);
+  });
+});
 
 //FUNCION CHEKEO PRODUCTOS EN LOCALSTORAGE
 
 function productosEnStorage () {
-  const itemsEnStorage = localStorage.getItem("NuevoItem")
-  console.log(itemsEnStorage)
+  const itemsEnStorage = localStorage.getItem("NuevoItem");
+  console.log(itemsEnStorage);
 
   if (itemsEnStorage != null){
     carrito = JSON.parse (itemsEnStorage);
     mostrarCarrito();
-  }
-}
+  };
+};
 
 // FUNCION DE BUSQUEDA DE PRODUCTOS REPETIDOS EN ARRAY CARRITO
 
 const buscarProducto = (idProd) => {
-  const objetoClickeado = inventario.find((elem) => elem.id === idProd)
+  const objetoClickeado = inventario.find((elem) => elem.id === idProd);
   const buscarEnCarrito = carrito.find ((e) => {return e.id === objetoClickeado.id});
 
   if (buscarEnCarrito === undefined){
@@ -72,27 +72,27 @@ const buscarProducto = (idProd) => {
   } else {
     objetoClickeado.cantidad +=1
     mostrarCarrito();
-  }
-}
+  };
+};
 
 // FUNCION CREAR CARRITO
 
 const mostrarCarrito = () => {
-  divCarrito.innerHTML = ""
+  divCarrito.innerHTML = "";
   carrito.forEach((elem) => {
 
-    const objetoenJSON = JSON.stringify (carrito)
-    localStorage.setItem ("NuevoItem", objetoenJSON)
+    const objetoenJSON = JSON.stringify (carrito);
+    localStorage.setItem ("NuevoItem", objetoenJSON);
 
-    total = carrito.reduce((accum,e) => {return accum += e.cantidad*e.precio}, 0)
+    total = carrito.reduce((accum,e) => {return accum += e.cantidad*e.precio}, 0);
 
     const divProductosCarrito = document.createElement('div')
     divProductosCarrito.innerHTML = `
     <p><u>${elem.nombre} </u></p>
     <p>- Precio: $${elem.precio} -</p>
-    <p>Cantidad: ${elem.cantidad} -</p>
+    <p>Cantidad:<button class="btn btn-black" onClick="">➖</button> ${elem.cantidad}<button class="btn btn-black" onClick="">✖️</button>  <hr> -</p>
     <p class="subtotal"><b>Subtotal: ${elem.precio * elem.cantidad}</b></p>
-    <button class="btn btn-black" onClick="borrarProducto()">❌</button>  <hr>`
+    <button class="btn btn-black" onClick="borrarProducto(${elem.id})">❌</button>  <hr>`
     ;
 
     divCarrito.appendChild(divProductosCarrito);
@@ -100,26 +100,66 @@ const mostrarCarrito = () => {
 
   divTotalCarrito.innerHTML = `
   <p><b> TOTAL: ${total} <b/></p>
-  <button class="btn btn-primary" onClick="borrarCarrito()">Borrar Carrito</button>`;
+  <button class="btn btn-primary" onClick="borrarCarrito()">Borrar Carrito</button>
+  <button class="btn btn-success" onClick="confirmarCompra()">Confirmar Compra</button>`;
 
-}
+};
 
 //llAMADO A FUNCION DE CHEKEO EN LOCALSTORAGE
 
 productosEnStorage()
 
-// FUNCION BORRAR CARRITO ENTERO
+// FUNCION BORRAR CARRITO ENTERO      // SWEETALERT 
 
 function borrarCarrito(){
-  total = 0
-  divCarrito.innerHTML= "";
-  divTotalCarrito.innerHTML = "";
-  carrito.splice(0,carrito.length);
-  localStorage.clear();
-}
+  Swal.fire({
+    title: '¿Estas seguro que quieres borrar el carrito?',
+    text: "",
+    icon: 'warning',
+    showCancelButton: true,
+    cancelButtonText: 'Cancelar',
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: '¡Borrar!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        '¡Carrito borrado!'
+      )
+      total = 0;
+      divCarrito.innerHTML= "";
+      divTotalCarrito.innerHTML = "";
+      carrito.splice(0,carrito.length);
+      localStorage.clear();
+    }
+  })
+};
 
 //FUNCION BORRAR ELEMENTO
 
-const borrarProducto = () => {
-  
+const borrarProducto = (idProd) => {
+  const elementoAEliminar = carrito.find((e) => e.id === idProd);
+
+  carrito = carrito.filter ((elem) => { return elem !== elementoAEliminar  })
+  console.log(carrito);
+
+  mostrarCarrito();
 };
+
+// CONFIRMAR COMPRA 
+
+function confirmarCompra () {
+  Swal.fire({
+    title: 'MUCHAS GRACIAS POR SU COMPRA ❤️',
+    width: 600,
+    padding: '3em',
+    color: '#716add',
+    background: '#fff url(/images/trees.png)',
+    backdrop: `
+      rgba(0,0,123,0.4)
+      url("/images/nyan-cat.gif")
+      left top
+      no-repeat
+    `
+  })
+}
